Attempt to fix cryppo mocking for tests that use cryppo
