+++ packages/cli/package.json
@@ -51,6 +51,7 @@
     "nyc": "~15.0.0",
     "prettier": "~1.19.1",
     "sinon": "~8.1.1",
+    "ts-mock-imports": "^1.3.0",
     "ts-node": "^8.1.0",
     "tsconfig-paths": "^3.9.0",
     "tslint": "^5.16.0",
