+++ packages/sdk/package.json
@@ -42,6 +42,7 @@
     "prettier": "~1.19.1",
     "rollup-plugin-typescript2": "^0.27.1",
     "sinon": "~8.1.1",
+    "ts-mock-imports": "^1.3.0",
     "ts-node": "^8.1.0",
     "tsconfig-paths": "^3.9.0",
     "tslint": "^5.16.0",
