+++ packages/sdk/src/services/cryppo-factory.ts
@@ -1,0 +1,5 @@
+import * as _cryppo from '@meeco/cryppo';
+
+export const cryppoFactory = () => {
+  return (<any>global).cryppo || _cryppo;
+};
